Interface web pour les scores
Ajout des pages web pour décrire le projet, ainsi que le classement des
scores enregistrés dans la base de donnée.

Modifications des classes InputKey et Bouton : le positionnement de
texte se fait maintenant au pixel près, évitant le flou correspondant au
décalage d'un nombre de pixel non entière, du texte. (en gros, le texte
n'est plus flou)

Le score affiché dans l'écran de jeu dans la case "Partie précédente"
n'affiche plus les parties à 0 point.

Le fichier web/include/id_bdd.php doit exister et contenir les
identifiants de connexion à la base de donnée (se baser sur
id_bdd.base.php).

+++ Cpp/Bouton.cpp
@@ -84,10 +84,10 @@ void Bouton::updateTextPosition()
 {
     texte.setOrigin(0, 0);
     texte.setPosition(0, 0);
-    texte.setOrigin(texte.getGlobalBounds().left + texte.getGlobalBounds().width/2.0,
-                    texte.getGlobalBounds().top + texte.getGlobalBounds().height/2.0);
-    texte.setPosition(_position.x+_size.x/2.0,
-                      _position.y+_size.y/2.0);
+    texte.setOrigin((int) (texte.getGlobalBounds().left + texte.getGlobalBounds().width/2.0),
+                    (int) (texte.getGlobalBounds().top + texte.getGlobalBounds().height/2.0));
+    texte.setPosition((int)(_position.x+_size.x/2.0),
+                      (int)(_position.y+_size.y/2.0));
 }
